+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Under development using Godot 4.5.1. We'll be moving on to 4.6 betas soon! Every
 * Added optional "stroboscope" properties in IVCoreSettings. These create an artificial stroboscope effect for fast rotating bodies that is more stable and pleasing then the effect you might see from frame updates. 
 
 ### Changed
+* Set root Window.mode = MODE_WINDOWED on OS.shell_open(url) call by default.
 * Changed default game speeds and speed names.
 * [API breaking] Removed property IVSpeedManager.speed_name. Use method IVSpeedManager.get_speed_name()
 * [API breaking] Removed IVGlobal.speeds and changed IVGlobal.times indexing.